chore: fix `func generateBlockWithRandomTrie` is unused
