+++ dot/state/service_integration_test.go
@@ -15,6 +15,7 @@ import (
 	"github.com/ChainSafe/gossamer/dot/types"
 	"github.com/ChainSafe/gossamer/internal/log"
 	"github.com/ChainSafe/gossamer/lib/common"
+	runtime "github.com/ChainSafe/gossamer/lib/runtime/storage"
 	"github.com/ChainSafe/gossamer/lib/trie"
 	"github.com/golang/mock/gomock"
 
@@ -425,4 +426,39 @@ func TestService_Import(t *testing.T) {
 
 	err = serv.Stop()
 	require.NoError(t, err)
+}
+
+func generateBlockWithRandomTrie(t *testing.T, serv *Service,
+	parent *common.Hash, bNum uint) (*types.Block, *runtime.TrieState) {
+	trieState, err := serv.Storage.TrieState(nil)
+	require.NoError(t, err)
+
+	// Generate random data for trie state.
+	rand := time.Now().UnixNano()
+	key := []byte("testKey" + fmt.Sprint(rand))
+	value := []byte("testValue" + fmt.Sprint(rand))
+	err = trieState.Put(key, value)
+	require.NoError(t, err)
+
+	trieStateRoot, err := trieState.Root()
+	require.NoError(t, err)
+
+	if parent == nil {
+		bb := serv.Block.BestBlockHash()
+		parent = &bb
+	}
+
+	body, err := types.NewBodyFromBytes([]byte{})
+	require.NoError(t, err)
+
+	block := &types.Block{
+		Header: types.Header{
+			ParentHash: *parent,
+			Number:     bNum,
+			StateRoot:  trieStateRoot,
+			Digest:     createPrimaryBABEDigest(t),
+		},
+		Body: *body,
+	}
+	return block, trieState
 }
