+++ dot/state/test_helpers.go
@@ -5,7 +5,6 @@ package state
 
 import (
 	"crypto/rand"
-	"fmt"
 	"math/big"
 	"testing"
 	"time"
@@ -13,7 +12,6 @@ import (
 	"github.com/ChainSafe/chaindb"
 	"github.com/ChainSafe/gossamer/dot/types"
 	"github.com/ChainSafe/gossamer/lib/common"
-	runtime "github.com/ChainSafe/gossamer/lib/runtime/storage"
 	"github.com/ChainSafe/gossamer/lib/trie"
 	"github.com/ChainSafe/gossamer/lib/utils"
 	"github.com/ChainSafe/gossamer/pkg/scale"
@@ -233,39 +231,4 @@ func AddBlocksToStateWithFixedBranches(t *testing.T, blockState *BlockState, dep
 			arrivalTime = arrivalTime.Add(inc)
 		}
 	}
-}
-
-func generateBlockWithRandomTrie(t *testing.T, serv *Service,
-	parent *common.Hash, bNum uint) (*types.Block, *runtime.TrieState) {
-	trieState, err := serv.Storage.TrieState(nil)
-	require.NoError(t, err)
-
-	// Generate random data for trie state.
-	rand := time.Now().UnixNano()
-	key := []byte("testKey" + fmt.Sprint(rand))
-	value := []byte("testValue" + fmt.Sprint(rand))
-	err = trieState.Put(key, value)
-	require.NoError(t, err)
-
-	trieStateRoot, err := trieState.Root()
-	require.NoError(t, err)
-
-	if parent == nil {
-		bb := serv.Block.BestBlockHash()
-		parent = &bb
-	}
-
-	body, err := types.NewBodyFromBytes([]byte{})
-	require.NoError(t, err)
-
-	block := &types.Block{
-		Header: types.Header{
-			ParentHash: *parent,
-			Number:     bNum,
-			StateRoot:  trieStateRoot,
-			Digest:     createPrimaryBABEDigest(t),
-		},
-		Body: *body,
-	}
-	return block, trieState
 }
